requirements.txt
asgiref==3.5.2
Django==4.1.4
django-appconf==1.0.5
django-compressor==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.7
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
tzdata==2022.7
